Add tab with active systems to log modal
